--- a/src/database.hpp
+++ b/src/database.hpp
@@ -84,6 +84,9 @@ struct Database
     // A temporary list of the nodes that we actually used
     std::vector<value_t> used_nodes_list;
 
+    // A list of the OSM way IDs
+    std::vector<std::uint64_t> internal_to_external_way_id_map;
+
   private:
     // The character data for all strings
     std::vector<char> string_data;