--- a/src/extractor.hpp
+++ b/src/extractor.hpp
@@ -6,6 +6,9 @@
 #include "database.hpp"
 #include "types.hpp"
 
+#include <string>
+#include <unordered_set>
+
 /**
  * The handler for libosmium.  This class basically contains one callback that's called by
  * libosmium as it parses our OSM file.
@@ -45,4 +48,6 @@ struct Extractor final : osmium::handler::Handler
     // Internal reference to the db we're going to dump everything
     // into
     Database &db;
+
+    const std::unordered_set<std::string> valid_highways;
 };