--- a/test/basic/extractor.cpp
+++ b/test/basic/extractor.cpp
@@ -32,7 +32,7 @@ BOOST_AUTO_TEST_CASE(annotator_test_basic)
     Extractor extractor(buffer.c_str(), buffer.size(), "xml", db);
     BOOST_CHECK(db.rtree);
 
-    BOOST_CHECK_EQUAL(db.pair_way_map.size(), 4);
+    BOOST_CHECK_EQUAL(db.pair_way_map.size(), 2);
     BOOST_CHECK_EQUAL(db.external_internal_map.size(), 3);
     BOOST_CHECK_EQUAL(db.external_internal_map[101], 0);
     BOOST_CHECK_EQUAL(db.external_internal_map[202], 1);