--- a/test/bench.cpp
+++ b/test/bench.cpp
@@ -1,4 +1,5 @@
 #include <cstdlib>
+#include <chrono>
 #include <iostream>
 #include <string>
 #include <utility>