--- a/test/bench.cpp
+++ b/test/bench.cpp
@@ -1,5 +1,4 @@
 #include <cstdlib>
-#include <chrono>
 #include <iostream>
 #include <string>
 #include <utility>