--- a/test/bench.cpp
+++ b/test/bench.cpp
@@ -1,3 +1,4 @@
+#include <chrono>
 #include <cstdlib>
 #include <iostream>
 #include <string>