--- a/test/bench.cpp
+++ b/test/bench.cpp
@@ -1,5 +1,6 @@
 #include <chrono>
 #include <cstdlib>
+#include <chrono>
 #include <iostream>
 #include <string>
 #include <utility>