This change implements a few speedups:

  - nodes are never examined if the rtree is not going to be built
  - use faster method for checking tags (only a small gain here, probably redundant in the face of https://github.com/mapbox/route-annotator/pull/25 anyway)
  - only store node pairs once, but flag direction

I tested this with `british-isles-latest.osm.pbf` which is about 1,201,699,727 bytes.  Before this change, loading the file took about 80 seconds.  After these three changes, load time is about 27 seconds, and memory usage is also decreased.  There should be little or no impact on query performance.

## Note on pre-filtering

If I'm just intersted in ways with `highway=*` and `maxspeed=*`, I can pre-filter the file for even more speedup:

```
osmium tags-filter  british-isles-latest.osm.pbf 'maxspeed' 'maxspeed:forward' 'maxspeed:backward' -o only_maxspeed.pbf
osmium tags-filter only_maxspeed.pbf highway -o only_maxspeed_highways.osm.pbf
```

This results in `only_maxspeed_highways.osm.pbf` that's around 94,376,696 bytes.

Running the benchmark code against this reduced file results in a load time of around 4 seconds, down from the original 80 seconds.